requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
psycopg[binary]==3.2.1
django-otp==1.3.0
django-two-factor-auth==1.15.5
django-cors-headers==4.3.1
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
drf-spectacular==0.27.2
weasyprint==60.2
azure-storage-blob==12.20.0
boto3==1.34.135
python-magic==0.4.27
Pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.2
PyJWT==2.9.0
django-tenants==3.6.1
gunicorn==22.0.0
flower==2.0.1
qrcode==7.4.2
stripe==10.8.0

# Testing dependencies
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0
faker==19.6.2

# Code quality
black==23.7.0
flake8==6.0.0
flake8-django==1.4.0
isort==5.12.0

# Security
bandit==1.7.5
safety==2.3.5
